Split app key from viewport title to stop spraying app.ron files everywhere (#59)
Egui uses the app key to determine the storage location for persisted
configuration files. This does not work well when the app key is
something like "~/Downloads/prof_0.gz - Legion Prof" (the case when
profiling files you've just downloaded), which results in the app.ron
files being persisted to some random subdirectory of ~/Downloads
instead of "~/Library/Application Support/Legion Prof" or something
similar (depending on the OS).

So split the app key off from the viewport title so that the former
can be short and predictable and not mess with the persistent config
location so badly.